Implement opening of exported PDF with default app
This commit implements the automatic opening of exported
PDF documents with the system default app. It can be
toggled via a checkbox in the export dialog.

NOTE: This is temporary code, since I am not too happy
with the `PDFDocument.open` method. I'll have to think
about that for a while first.

Squashed commits:
- Ignore node_modules/
- Open PDF document after export
- Implement UI for toggling opening of PDF documents