Closing #5: Added Gruntfile with all necessary tasks to prepare builds
In order to do this, apart fron the Gruntfile.js file, the package.json file had to be changed as well and includes a number of devDependencies. Another small change was done in the extension name, title and I added the repository field in the file to remove the warning about this field.